+++ source/standalone/workflows/skrl_ctrl/networks/actor.py
@@ -183,3 +183,128 @@ class StochasticActor(GaussianMixin, Model):
 
     def compute(self, inputs, role):
         return self.net(inputs["states"]), self.log_std_parameter, {}
+      
+
+
+
+
+
+class TanhTransform2(pyd.transforms.Transform):
+  domain = pyd.constraints.real
+  codomain = pyd.constraints.interval(-1.0, 1.0)
+  bijective = True
+  sign = +1
+
+  def __init__(self, cache_size=1):
+    super().__init__(cache_size=cache_size)
+
+  @staticmethod
+  def atanh(x):
+    return 0.5 * (x.log1p() - (-x).log1p())
+
+  def __eq__(self, other):
+    return isinstance(other, TanhTransform2)
+
+  def _call(self, x):
+    return x.tanh()
+
+  def _inverse(self, y):
+    # We do not clamp to the boundary here as it may degrade the performance of certain algorithms.
+    # one should use `cache_size=1` instead
+    return self.atanh(y)
+
+  def log_abs_det_jacobian(self, x, y):
+    # We use a formula that is more numerically stable, see details in the following link
+    # https://github.com/tensorflow/probability/commit/ef6bb176e0ebd1cf6e25c6b5cecdd2428c22963f#diff-e120f70e92e6741bca649f04fcd907b7
+    return 2. * (math.log(2.) - x - F.softplus(-2. * x))
+
+
+class SquashedNormal2(pyd.transformed_distribution.TransformedDistribution):
+  def __init__(self, loc, scale):
+    self.loc = loc
+    self.scale = scale
+
+    self.base_dist = pyd.Normal(loc, scale)
+    transforms = [TanhTransform2()]
+    super().__init__(self.base_dist, transforms)
+
+  @property
+  def mean(self):
+    mu = self.loc
+    for tr in self.transforms:
+        mu = tr(mu)
+    return mu
+
+
+class DiagGaussianActor2(nn.Module):
+  """torch.distributions implementation of an diagonal Gaussian policy."""
+  def __init__(self, obs_dim, action_dim, hidden_dim, hidden_depth,
+                log_std_bounds):
+    super().__init__()
+
+    self.log_std_bounds = log_std_bounds
+    self.trunk = mlp(obs_dim, hidden_dim, 2 * action_dim,
+                            hidden_depth)
+
+    self.outputs = dict()
+    self.apply(weight_init)
+
+  def forward(self, obs):
+    mu, log_std = self.trunk(obs).chunk(2, dim=-1)
+
+    # constrain log_std inside [log_std_min, log_std_max]
+    log_std = torch.tanh(log_std)
+    log_std_min, log_std_max = self.log_std_bounds
+    log_std = log_std_min + 0.5 * (log_std_max - log_std_min) * (log_std +
+                                                                  1)
+
+    std = log_std.exp()
+
+    self.outputs['mu'] = mu
+    self.outputs['std'] = std
+
+    dist = SquashedNormal2(mu, std)
+    return dist
+
+
+class DiagGaussianActorPolicy(GaussianMixin, Model):
+  """torch.distributions implementation of an diagonal Gaussian policy."""
+  def __init__(self, observation_space, action_space, hidden_dim, hidden_depth,
+                log_std_bounds, device):
+    Model.__init__(self, observation_space, action_space, device)
+    GaussianMixin.__init__(self, min_log_std=log_std_bounds[0], max_log_std=log_std_bounds[1], reduction="mean")
+
+    # self.log_std_bounds = log_std_bounds
+    # # self.trunk = mlp(observation_space.shape[0], hidden_dim, 2 * action_space.shape[0],
+    # #                         hidden_depth)
+    # self.trunk = nn.Sequential(nn.Linear(self.num_observations, 1024),
+    #                              nn.ReLU(),
+    #                              nn.Linear(1024, 512),
+    #                              nn.ReLU(),
+    #                              nn.Linear(512, 2 * self.num_actions))
+    # self.apply(weight_init)
+    
+    self.actor = DiagGaussianActor2(obs_dim = observation_space.shape[0],
+                                    action_dim = action_space.shape[0],
+                                    hidden_dim = hidden_dim,
+                                    hidden_depth = hidden_depth,
+                                    log_std_bounds = log_std_bounds)
+    
+    
+  def compute(self, inputs, role):
+    # mu, log_std = self.trunk(inputs["states"]).chunk(2, dim=-1)
+
+    # # constrain log_std inside [log_std_min, log_std_max]
+    # log_std = torch.tanh(log_std)
+    # log_std_min, log_std_max = self.log_std_bounds
+    # log_std = log_std_min + 0.5 * (log_std_max - log_std_min) * (log_std +
+    #                                                               1)
+    
+    dist = self.actor(inputs['states'])
+    # sample using the reparameterization trick
+    actions = dist.rsample()
+    
+    # log of the probability density function
+    log_prob = dist.log_prob(actions).sum(-1, keepdim=True)
+    
+    return actions, log_prob, {}
