+++ source/standalone/workflows/skrl_ctrl/train.py
@@ -10,7 +10,7 @@ from skrl.resources.preprocessors.torch import RunningStandardScaler
 from skrl.trainers.torch import SequentialTrainer, StepTrainer
 from skrl.utils import set_seed
 
-from sac.actor import DiagGaussianActor, StochasticActor
+from sac.actor import DiagGaussianActor, StochasticActor, DiagGaussianActorPolicy
 from sac.critic import Critic, TestCritic
 from sac.feature import Phi, Mu, Theta
 
@@ -28,11 +28,12 @@ set_seed(42)  # e.g. `set_seed(42)` for fixed seed
 cli_args = ["--video"]
 # load and wrap the Isaac Gym environment
 task_version = "Multi"
+multitask = True
 task_name = f"Isaac-Quadcopter-{task_version}-Trajectory-Direct-v0"
-env = load_isaaclab_env(task_name = task_name, num_envs=64, cli_args=cli_args)
+env = load_isaaclab_env(task_name = task_name, num_envs=32, cli_args=cli_args)
 
 video_kwargs = {
-    "video_folder": os.path.join(f"runs/torch/{task_version}/", "videos", "train"),
+    "video_folder": os.path.join(f"runs/torch/{task_version}/{str(multitask)}", "videos", "train"),
     "step_trigger": lambda step: step % 10000== 0,
     "video_length": 400,
     "disable_logger": True,
@@ -53,7 +54,7 @@ memory = RandomMemory(memory_size=memory_size, num_envs=env.num_envs, device=dev
 
 # define hidden dimension
 actor_hidden_dim = 512
-actor_hidden_depth = 2
+actor_hidden_depth = 3
 
 # define feature dimension 
 feature_dim = 512
@@ -63,7 +64,7 @@ feature_hidden_dim = 1024
 # SAC requires 5 models, visit its documentation for more details
 # https://skrl.readthedocs.io/en/latest/api/agents/sac.html#models
 models = {}
-models["policy"] = DiagGaussianActor(observation_space = env.observation_space,
+models["policy"] = StochasticActor(observation_space = env.observation_space,
                                      action_space = env.action_space, 
                                      hidden_dim = actor_hidden_dim, 
                                      hidden_depth = actor_hidden_depth,
@@ -124,7 +125,7 @@ models["mu"] = Mu(
 # https://skrl.readthedocs.io/en/latest/api/agents/sac.html#configuration-and-hyperparameters
 cfg = SAC_DEFAULT_CONFIG.copy()
 cfg["gradient_steps"] = 1
-cfg["batch_size"] = 1024
+cfg["batch_size"] = 256
 cfg["discount_factor"] = 0.99
 cfg["polyak"] = 0.005
 cfg["actor_learning_rate"] = 1e-4
@@ -135,7 +136,7 @@ cfg["random_timesteps"] = 25e3
 cfg["learning_starts"] = 25e3
 cfg["grad_norm_clip"] = 1.0
 cfg["learn_entropy"] = True
-cfg["entropy_learning_rate"] = 1e-4
+cfg["entropy_learning_rate"] = 1e-5
 cfg["initial_entropy_value"] = 1.0
 # cfg["state_preprocessor"] = RunningStandardScaler
 # cfg["state_preprocessor_kwargs"] = {"size": env.observation_space, "device": device}
@@ -143,7 +144,8 @@ cfg["initial_entropy_value"] = 1.0
 cfg["experiment"]["write_interval"] = 1000
 cfg["experiment"]["checkpoint_interval"] = 10000
 cfg['use_feature_target'] = True
-cfg['extra_feature_steps'] = 1
+cfg['extra_feature_steps'] = 0
+cfg['extra_critic_steps'] = 2
 cfg['target_update_period'] = 1
 cfg['eval'] = False
 
@@ -159,7 +161,7 @@ agent = CTRLSACAgent(
             device=device
         )
 
-cfg_trainer = {"timesteps": int(5e5), "headless": True}
+cfg_trainer = {"timesteps": int(1e6), "headless": True}
 trainer = SequentialTrainer(cfg=cfg_trainer, env=env, agents=agent)
 
 # train the agent(s)
